0.6.0

+++ src/main/java/me/travis/wurstplusthree/mixin/mixins/MixinRenderPlayer.java
@@ -41,7 +41,7 @@ public class MixinRenderPlayer {
             GL11.glEnable((int)10754);
             OpenGlHelper.setLightmapTextureCoords((int)OpenGlHelper.lightmapTexUnit, (float)240.0f, (float)240.0f);
             Color color = HandColour.INSTANCE.colour.getValue();
-            GL11.glColor4f((float)((float)color.getRed() / 255.0f), (float)((float)color.getGreen() / 255.0f), (float)color.getBlue() / 255.0f,
+            GL11.glColor4f((float)color.getRed() / 255.0f, (float)((float)color.getGreen() / 255.0f), (float)color.getBlue() / 255.0f,
                     (float)color.getAlpha() / 255.0f);
         }
         else if(clientPlayer == Minecraft.getMinecraft().player && PlayerSpoofer.INSTANCE.isEnabled()){
